Turn Util module into a package
Change Util.py into a package.  Some routines are split out
into separate files in this package.
